Разбивает JS-код на модули
